fix: format string type mismatch for file_size on ARM platforms
Fixed compilation error on ARM (32-bit) platforms where st_size is 'long int'
instead of 'size_t':

- Cast file_size to size_t in LOG_INFO format string
- Resolves: format '%zu' expects argument of type 'size_t', but has type 'long int'

This is common on 32-bit ARM systems where file sizes are represented as long
instead of size_t. The cast ensures compatibility across all architectures.

+++ src/extensions/vfs_memory.c
@@ -367,7 +367,7 @@ static int vfs_load_file_from_disk(const char *file_path, const char *base_path)
 			                       &decrypted, &decrypted_len) == 0) {
 				/* Decryption successful */
 				LOG_INFO("VFS: Successfully decrypted file: %s (%zu -> %zu bytes)",
-				         file_path, file_size, decrypted_len);
+				         file_path, (size_t)file_size, decrypted_len);
 				
 				/* Replace content with decrypted data */
 				free(content);
